YAYY

+++ The Red Room/Assets/TeleportationScript.cs
@@ -8,7 +8,7 @@ public class TeleportationScript : MonoBehaviour
 
     private void OnGUI()
     {
-        if(GUI.Button(new Rect(-884.438, -438.829, -57.7, 95.61768), "BasePlate"))
+        if(GUI.Button(new Rect((float)-884.438, (float) -438.829, (float) -57.7, (float)95.61768), "BasePlate"))
         {
             transform.position = Teleport.position;
         }
